I finished the data cleaning for Q2.
"Categorical variables will be ordered alphabetically; factors will follow the specified order level that underlies the variable labels. You can change the order level of a factor variable to your specified preference using forcats::fct_relevel or according to the value of another variable using forcats::fct_reorder." Visualization part 2 under data manipulation

Remember that %in% checks whether or not the object is contained in the other object.  == is a logical operator that checks for identity properties.